Split non commuting
