+++ tests/legacy_return/test_split_non_commuting.py
@@ -139,80 +139,29 @@ class TestUnittestSplitNonCommuting:
             split_non_commuting(tape)
 
 
-class TestIntegration:
-    """Integration tests for ``qml.transforms.split_non_commuting()``"""
+# Integration test
+def test_expval_non_commuting_observables():
+    """Test expval with multiple non-commuting operators"""
+    dev = qml.device("default.qubit", wires=6)
 
-    def test_expval_non_commuting_observables(self):
-        """Test expval with multiple non-commuting operators"""
-        dev = qml.device("default.qubit", wires=6)
+    @qml.qnode(dev)
+    def circuit():
+        qml.Hadamard(1)
+        qml.Hadamard(0)
+        qml.PauliZ(0)
+        qml.Hadamard(3)
+        qml.Hadamard(5)
+        qml.T(5)
+        return [
+            qml.expval(qml.PauliZ(0) @ qml.PauliZ(1)),
+            qml.expval(qml.PauliX(0)),
+            qml.expval(qml.PauliZ(1)),
+            qml.expval(qml.PauliX(1) @ qml.PauliX(4)),
+            qml.expval(qml.PauliX(3)),
+            qml.expval(qml.PauliY(5)),
+        ]
 
-        @qml.qnode(dev)
-        def circuit():
-            qml.Hadamard(1)
-            qml.Hadamard(0)
-            qml.PauliZ(0)
-            qml.Hadamard(3)
-            qml.Hadamard(5)
-            qml.T(5)
-            return (
-                qml.expval(qml.PauliZ(0) @ qml.PauliZ(1)),
-                qml.expval(qml.PauliX(0)),
-                qml.expval(qml.PauliZ(1)),
-                qml.expval(qml.PauliX(1) @ qml.PauliX(4)),
-                qml.expval(qml.PauliX(3)),
-                qml.expval(qml.PauliY(5)),
-            )
-
-        res = circuit()
-        assert isinstance(res, tuple)
-        assert len(res) == 6
-        assert all(isinstance(r, np.ndarray) for r in res)
-        assert all(r.shape == () for r in res)
-
-        res = qml.math.stack(res)
-
-        assert all(np.isclose(res, np.array([0.0, -1.0, 0.0, 0.0, 1.0, 1 / np.sqrt(2)])))
-
-    def test_shot_vector_support(self):
-        """Test output is correct when using shot vectors"""
-
-        dev = qml.device("default.qubit", wires=6, shots=(10000, (20000, 2), 30000))
-
-        @qml.qnode(dev)
-        def circuit():
-            qml.Hadamard(1)
-            qml.Hadamard(0)
-            qml.PauliZ(0)
-            qml.Hadamard(3)
-            qml.Hadamard(5)
-            qml.T(5)
-            return (
-                qml.expval(qml.PauliZ(0) @ qml.PauliZ(1)),
-                qml.expval(qml.PauliX(0)),
-                qml.expval(qml.PauliZ(1)),
-                qml.expval(
-                    qml.PauliY(0) @ qml.PauliY(1) @ qml.PauliZ(3) @ qml.PauliY(4) @ qml.PauliX(5)
-                ),
-                qml.expval(qml.PauliX(1) @ qml.PauliX(4)),
-                qml.expval(qml.PauliX(3)),
-                qml.expval(qml.PauliY(5)),
-            )
-
-        res = circuit()
-        assert isinstance(res, tuple)
-        assert len(res) == 4
-        assert all(isinstance(shot_res, tuple) for shot_res in res)
-        assert all(len(shot_res) == 7 for shot_res in res)
-        assert all(
-            all(list(isinstance(r, np.ndarray) and r.shape == () for r in shot_res))
-            for shot_res in res
-        )
-
-        res = qml.math.stack([qml.math.stack(r) for r in res])
-
-        assert np.allclose(
-            res, np.array([0.0, -1.0, 0.0, 0.0, 0.0, 1.0, 1 / np.sqrt(2)]), atol=0.05
-        )
+    assert all(np.isclose(circuit(), np.array([0.0, -1.0, 0.0, 0.0, 1.0, 1 / np.sqrt(2)])))
 
 
 # Autodiff tests
@@ -240,16 +189,13 @@ class TestAutodiffSplitNonCommuting:
                 qml.expval(qml.PauliZ(0)),
             )
 
-        def cost(params):
-            res = circuit(params)
-            return qml.math.stack(res)
-
         params = pnp.array([0.5, 0.5])
-        res = cost(params)
-        grad = qml.jacobian(cost)(params)
+        res = circuit(params)
+        grad = qml.jacobian(circuit)(params)
         assert all(np.isclose(res, exp_res))
         assert all(np.isclose(grad, exp_grad).flatten())
 
+    # TODO: Currently not possible to jit multiple expvals
     @pytest.mark.jax
     def test_split_with_jax(self):
         """Test that results after splitting are still differentiable with jax"""
@@ -259,70 +205,6 @@ class TestAutodiffSplitNonCommuting:
 
         dev = qml.device("default.qubit.jax", wires=3)
 
-        @qml.qnode(dev)
-        def circuit(params):
-            qml.RX(params[0], wires=0)
-            qml.RY(params[1], wires=1)
-            return (
-                qml.expval(qml.PauliZ(0) @ qml.PauliZ(1)),
-                qml.expval(qml.PauliY(0)),
-                qml.expval(qml.PauliZ(0)),
-            )
-
-        params = jnp.array([0.5, 0.5])
-        res = circuit(params)
-        grad = jax.jacobian(circuit)(params)
-        assert all(np.isclose(res, exp_res))
-        assert all(np.isclose(grad, exp_grad, atol=1e-5).flatten())
-
-    @pytest.mark.jax
-    def test_split_with_jax_multi_params(self):
-        """Test that results after splitting are still differentiable with jax
-        with multiple parameters"""
-
-        import jax
-        import jax.numpy as jnp
-
-        dev = qml.device("default.qubit.jax", wires=3)
-
-        @qml.qnode(dev)
-        def circuit(x, y):
-            qml.RX(x, wires=0)
-            qml.RY(y, wires=1)
-            return (
-                qml.expval(qml.PauliZ(0) @ qml.PauliZ(1)),
-                qml.expval(qml.PauliY(0)),
-                qml.expval(qml.PauliZ(0)),
-            )
-
-        x = jnp.array(0.5)
-        y = jnp.array(0.5)
-
-        res = circuit(x, y)
-        grad = jax.jacobian(circuit, argnums=[0, 1])(x, y)
-
-        assert all(np.isclose(res, exp_res))
-
-        assert isinstance(grad, tuple)
-        assert len(grad) == 3
-
-        for i, meas_grad in enumerate(grad):
-            assert isinstance(meas_grad, tuple)
-            assert len(meas_grad) == 2
-            assert all(isinstance(g, jnp.ndarray) and g.shape == () for g in meas_grad)
-
-            assert np.allclose(meas_grad, exp_grad[i], atol=1e-5)
-
-    @pytest.mark.jax
-    def test_split_with_jax_jit(self):
-        """Test that results after splitting are still differentiable with jax-jit"""
-
-        import jax
-        import jax.numpy as jnp
-
-        dev = qml.device("default.qubit", wires=3)
-
-        @jax.jit
         @qml.qnode(dev)
         def circuit(params):
             qml.RX(params[0], wires=0)
@@ -358,13 +240,9 @@ class TestAutodiffSplitNonCommuting:
                 qml.expval(qml.PauliZ(0)),
             )
 
-        def cost(params):
-            res = circuit(params)
-            return qml.math.stack(res)
-
         params = torch.tensor([0.5, 0.5], requires_grad=True)
-        res = cost(params)
-        grad = jacobian(cost, (params))
+        res = circuit(params)
+        grad = jacobian(circuit, (params))
         assert all(np.isclose(res.detach().numpy(), exp_res))
         assert all(np.isclose(grad.detach().numpy(), exp_grad, atol=1e-5).flatten())
 
@@ -390,7 +268,6 @@ class TestAutodiffSplitNonCommuting:
         res = circuit(params)
         with tf.GradientTape() as tape:
             loss = circuit(params)
-            loss = tf.stack(loss)
 
         grad = tape.jacobian(loss, params)
         assert all(np.isclose(res, exp_res))
